requirements.txt
flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
openai==0.27.0
beautifulsoup4==4.9.3
nltk==3.6.3
pandas==1.3.3
scikit-learn==0.24.2
transformers==4.30.0
huggingface-hub==0.16.4
torch==2.0.1
numpy==1.24.3
tqdm==4.65.0
newsapi-python==0.2.6
python-whois==0.7.3 